Move BLUESPAWN to BLUESPAWN-agent
Long term, there will be multiple seperate programs that are part of a
larger BLUESPAWN network -- the agent, runing on endpoints, a server
which orchestrates multiple agents on the same network, and a cloud
which functions as a public database of malware signatures gathered by
the agents.

This renaming is in preperation to implementing these other related
programs with an intuitive project structure.